+++ payment-engine/src/main/java/io/gomobi/payment/repository/SubMerchantRepository.java
@@ -3,9 +3,10 @@ package io.gomobi.payment.repository;
 import io.gomobi.payment.entity.SubMerchant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface SubMerchantRepository extends JpaRepository<SubMerchant, Long> {
+public interface SubMerchantRepository extends JpaRepository<SubMerchant, BigInteger> {
 
     Optional<SubMerchant> findBySubMerchantMid(String subMerchantMid);
 }
